:+1: Use new LatexSourceGenerator
